Refactor: card service interface와 Impl 분리, converter 클래스 구성
